feat: more great changes

+++ packages/contracts/test/systems/OverrideSystem.t.sol
@@ -151,7 +151,7 @@ contract OverrideSystemTest is PrimodiumTest {
     world.Empires__createShip{ value: totalCost }(planetId, 1);
     assertGt(LibPrice.getTotalCost(EOverride.CreateShip, empire, 1), totalCost, "Total Cost should have increased");
     assertGt(OverrideCost.get(empire, EOverride.CreateShip), overrideCost, "Override Cost should have increased");
-    assertEq(Player.getSpent(aliceId), totalCost, "Player should have spent total cost");
+    assertEq(uint256(Player.getSpent(aliceId)), totalCost, "Player should have spent total cost");
     assertEq(Balances.get(EMPIRES_NAMESPACE_ID), totalCost, "Namespace should have received the balance");
     assertEq(
       PointsMap.getValue(EEmpire.Red, aliceId),
@@ -174,7 +174,7 @@ contract OverrideSystemTest is PrimodiumTest {
       "Total Cost should have increased"
     );
     assertGt(OverrideCost.get(empire, EOverride.CreateShip), overrideCost, "Override Cost should have increased");
-    assertEq(Player.getSpent(aliceId), totalCost, "Player should have spent total cost");
+    assertEq(uint256(Player.getSpent(aliceId)), totalCost, "Player should have spent total cost");
     assertEq(Balances.get(EMPIRES_NAMESPACE_ID), totalCost, "Namespace should have received the balance");
     assertEq(
       PointsMap.getValue(EEmpire.Red, aliceId),
@@ -195,7 +195,7 @@ contract OverrideSystemTest is PrimodiumTest {
     world.Empires__killShip{ value: totalCost }(planetId, 1);
     assertGt(LibPrice.getTotalCost(EOverride.KillShip, empire, 1), totalCost, "Total Cost should have increased");
     assertGt(OverrideCost.get(empire, EOverride.KillShip), overrideCost, "Override Cost should have increased");
-    assertEq(Player.getSpent(bobId), totalCost, "Player should have spent total cost");
+    assertEq(uint256(Player.getSpent(bobId)), totalCost, "Player should have spent total cost");
     assertEq(Balances.get(EMPIRES_NAMESPACE_ID), initBalance + totalCost, "Namespace should have received the balance");
     assertEq(PointsMap.getValue(EEmpire.Blue, bobId), pointUnit, "Player should have received blue points");
     assertEq(PointsMap.getValue(EEmpire.Green, bobId), pointUnit, "Player should have received green points");
@@ -218,7 +218,7 @@ contract OverrideSystemTest is PrimodiumTest {
       "Total Cost should have increased"
     );
     assertGt(OverrideCost.get(empire, EOverride.KillShip), overrideCost, "Override Cost should have increased");
-    assertEq(Player.getSpent(bobId), totalCost, "Player should have spent total cost");
+    assertEq(uint256(Player.getSpent(bobId)), totalCost, "Player should have spent total cost");
     assertEq(Balances.get(EMPIRES_NAMESPACE_ID), initBalance + totalCost, "Namespace should have received the balance");
     assertEq(
       PointsMap.getValue(EEmpire.Blue, bobId),
